Add explicit param name arguments to helpers
The helpers used the field names before, which was more convenient for
very minimal cases, but for more complex queries (like when you have
multiple comparisons to a field) it'll lead to unintentional parameter conflicts.

If rust had an idiom default arguments that would be best, but for now
more flexible helpers is good I think and still saves plenty of typing... coming up with parameter names shouldn't be a burden either.